elasticsearch-create-backup-repository.sh: Fix verbosity of curl
Curl was to verbose, so even when succeeding the cron would send us an
email to inform us. That is unecessary and also spam for no apparent
reasons. After applying this commit, the crons of ES backups should
notify us only when something wrong happened - which is the expected
behavior.