mongodump: Fixing unecessary output when dumping
The output of the command provoques the trigger of email by cron because
it thinks the program had an issue - despite a proper return code.
The mongodump has been silenced in order to avoid unecessary
email/output. If we recieve and email from this cron, we will now know
that something went wront.